Since PHP 7.0, functions inspecting arguments no longer report runtime values passed.
This means that when function arguments are altered inside the function,
those functions (like func_get_args or debug_backtrace) no longer report
the correct value the function was called with.

Due to this change, Lang::set_lang() will not reload all language files
in the new language set, but in the original language determined when it
was first called, due to the change in runtime value in line 65.